preprocess/data_agent/requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0

# Data processing dependencies
pandas>=2.0.0
pyarrow>=14.0.0
tqdm>=4.66.0

# Optional dependencies
pydantic>=2.0.0
typing-extensions>=4.0.0

# Trading backend dependencies
backtrader>=1.9.78.123
ccxt
